Optimize AsyncClientManager by moving hashing to control plane (#29199)
This change removes the hashing of the config away from the data plane on every request to the control plane when the filter callback is created. This is accomplished by creating a wrapper object for the config which precomputes the hash and also serializes it to string ahead of time so that calls to the AsyncClientManager with that wrapper can avoid unnecessary overhead.

Signed-off-by: AlanDiaz <diazalan@google.com>