Cherry-pick c745c1956cdaffa35d6f4933b57a431a2d2e9fc7
The above commit is the changes that we needed from upstream envoy to support
using a typedExtensionConfig object as the output sink. Similar comments to
https://github.com/solo-io/envoy-gloo/commit/e85d08fa3f0232f73890e328fddf5e303ef16811

Prior to this cherry-pick, I was just applying a patch in `envoy-gloo`, but
this proved untenable with CI, so I'm bringing it into this branch to see if we
can make things work again without having to hack around with patches.
Additionally, this is an attempt to start reconciling our divergences from the
ext_proc work. I haven't tested any ext_proc changes here, but I hope that at
least by starting off with this envoy version, we can also make our updates in
gloo easier as well.